py_backend/requirements.txt
numpy==1.16.2
Flask_SQLAlchemy==2.4.0
Werkzeug==0.14.1
Flask==1.0.2
opencv_python==4.1.0.25
Flask_Session==0.3.1
Flask_SocketIO==3.3.2
typing==3.6.2
pycrypto==2.6.1
